Move creation of agent policy to project
