+++ internal/stack/serverless.go
@@ -6,12 +6,8 @@ package stack
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
 	"strings"
 	"time"
 
@@ -53,11 +49,12 @@ type projectSettings struct {
 	StackVersion string
 }
 
-func (sp *serverlessProvider) createProject(settings projectSettings, options Options) (Config, error) {
+func (sp *serverlessProvider) createProject(settings projectSettings, options Options, conf Config) (Config, error) {
 	project, err := sp.client.CreateProject(settings.Name, settings.Region, settings.Type)
 	if err != nil {
 		return Config{}, fmt.Errorf("failed to create %s project %s in %s: %w", settings.Type, settings.Name, settings.Region, err)
 	}
+	// project, _ := sp.currentProject(conf)
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Minute*30)
 	defer cancel()
@@ -166,19 +163,19 @@ func (sp *serverlessProvider) currentProject(config Config) (*serverless.Project
 	project.Credentials.Username = config.ElasticsearchUsername
 	project.Credentials.Password = config.ElasticsearchPassword
 
-	fleetURL, ok := config.Parameters[paramServerlessFleetURL]
-	if !ok {
+	project, err = sp.createClients(project)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create project client")
+	}
+
+	fleetURL := config.Parameters[paramServerlessFleetURL]
+	if true {
 		fleetURL, err = project.DefaultFleetServerURL(context.Background())
 		if err != nil {
 			return nil, fmt.Errorf("failed to get fleet URL: %w", err)
 		}
 	}
 	project.Endpoints.Fleet = fleetURL
-
-	project, err = sp.createClients(project)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create project client")
-	}
 
 	return project, nil
 }
@@ -228,13 +225,18 @@ func (sp *serverlessProvider) BootUp(options Options) error {
 		return err
 	case errProjectNotExist:
 		logger.Infof("Creating project %q", settings.Name)
-		config, err = sp.createProject(settings, options)
+		config, err = sp.createProject(settings, options, config)
 		if err != nil {
 			return fmt.Errorf("failed to create deployment: %w", err)
 		}
 
+		project, err = sp.currentProject(config)
+		if err != nil {
+			return fmt.Errorf("failed to retrieve latest project created: %w", err)
+		}
+
 		logger.Infof("Creating agent policy")
-		err = sp.createAgentPolicy(config, options.StackVersion)
+		err = project.CreateAgentPolicy(options.StackVersion)
 		if err != nil {
 			return fmt.Errorf("failed to create agent policy: %w", err)
 		}
@@ -294,80 +296,6 @@ func (sp *serverlessProvider) startLocalAgent(options Options, config Config) er
 	}
 
 	return nil
-}
-
-func (sp *serverlessProvider) createAgentPolicy(config Config, stackVersion string) error {
-	systemVersion, err := getPackageVersion("https://epr.elastic.co", "system", stackVersion)
-	if err != nil {
-		return fmt.Errorf("could not get the system package version for kibana %v: %w", stackVersion, err)
-	}
-
-	client, err := NewKibanaClient()
-	if err != nil {
-		return fmt.Errorf("failed to create kibana client: %w", err)
-	}
-
-	policy := kibana.Policy{
-		ID:                "elastic-agent-managed-ep",
-		Name:              "Elastic-Agent (elastic-package)",
-		Description:       "Policy created by elastic-package",
-		Namespace:         "default",
-		MonitoringEnabled: []string{"logs", "metrics"},
-	}
-	newPolicy, err := client.CreatePolicy(policy)
-	if err != nil {
-		return fmt.Errorf("error while creating agent policy: %w", err)
-	}
-
-	packagePolicy := kibana.PackagePolicy{
-		Name:     "system-1",
-		PolicyID: newPolicy.ID,
-	}
-	packagePolicy.Package.Name = "system"
-	packagePolicy.Package.Version = systemVersion
-
-	_, err = client.CreatePackagePolicy(packagePolicy)
-	if err != nil {
-		return fmt.Errorf("error while creating package policy: %w", err)
-	}
-
-	return nil
-}
-
-func getPackageVersion(registryURL, packageName, stackVersion string) (string, error) {
-	searchURL, err := url.JoinPath(registryURL, "search")
-	if err != nil {
-		return "", fmt.Errorf("could not build URL: %w", err)
-	}
-	searchURL = fmt.Sprintf("%s?package=%s&kibana.version=%s", searchURL, packageName, stackVersion)
-	resp, err := http.Get(searchURL)
-	if err != nil {
-		return "", fmt.Errorf("request failed (url: %s): %w", searchURL, err)
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode >= 300 {
-		return "", fmt.Errorf("unexpected status code %v", resp.StatusCode)
-	}
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", fmt.Errorf("failed to read response body: %w", err)
-	}
-	var packages []struct {
-		Name    string `json:"name"`
-		Version string `json:"version"`
-	}
-	err = json.Unmarshal(body, &packages)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse response body: %w", err)
-	}
-	if len(packages) != 1 {
-		return "", fmt.Errorf("expected 1 package, obtained %v", len(packages))
-	}
-	if found := packages[0].Name; found != packageName {
-		return "", fmt.Errorf("expected package %s, found %s", packageName, found)
-	}
-
-	return packages[0].Version, nil
 }
 
 func (sp *serverlessProvider) TearDown(options Options) error {
